Created getPage, getOffer, validate methods to ProfitCRUD class, fixed some old methods, added SQL script for creating schema of DB

+++ ProfitApp/src/main/java/bsu/fpmi/profit/Ad.java
@@ -14,12 +14,12 @@ public class Ad {
     private List<String> hashTags;
     private String discount;
     private Date validUntil;
-    private int rating;
+    private double rating;
     private List<String> reviews;
 
     public Ad(String id,String label, String description, String createdAt, String link,
               String vendor, String photoLink, List<String> hashTags,
-              String discount, String validUntil, int rating,
+              String discount, String validUntil, double rating,
               List<String> reviews){
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
         try {
@@ -77,7 +77,7 @@ public class Ad {
         return validUntil;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
@@ -136,7 +136,7 @@ public class Ad {
         this.validUntil = validUntil;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
